+++ test/try_catch.spec.ts
@@ -1,4 +1,4 @@
-import { tryCatch } from '../src/try_catch';
+import { isEither, tryCatch } from '../src';
 
 describe('tryCatch', () => {
   test('tryCatch function without catch function and with right case returns Either with Right state', () => {
@@ -27,35 +27,39 @@ describe('tryCatch', () => {
     expect(safeFn('error').extract()).toEqual(new Error('error'));
   });
 
-  test('tryCatch function with catch function and with right case returns Either with Right state', () => {
+  test('tryCatch function with catch function and with right case returns result', () => {
+    const text = 'input is not an error';
+
     const safeFn = tryCatch(
       (n) => {
         if (n === 'error') {
           throw new Error(n);
         } else {
-          return 'input is not an error';
+          return text;
         }
       },
       (error) => `Error: ${error.message}`
     );
 
-    expect(safeFn('default').isRight()).toBe(true);
-    expect(safeFn('default').extract()).toBe('input is not an error');
+    expect(isEither(safeFn('default'))).toBe(false);
+    expect(safeFn('default')).toBe(text);
   });
 
-  test('tryCatch function with catch function and with left case returns Either with Right state', () => {
+  test('tryCatch function with catch function and with left case returns result from catch function', () => {
+    const text = 'input is not an error';
+
     const safeFn = tryCatch(
       (n) => {
         if (n === 'error') {
           throw new Error(n);
         } else {
-          return 'input is not an error';
+          return text;
         }
       },
       (error) => `Error: ${error.message}`
     );
 
-    expect(safeFn('error').isLeft()).toBe(false);
-    expect(safeFn('error').extract()).toEqual('Error: error');
+    expect(isEither(safeFn('error'))).toBe(false);
+    expect(safeFn('error')).toEqual('Error: error');
   });
 });